plugin/requirements.txt
requests
mcp
python-dotenv